requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-seed==0.3.1
django-stubs[compatible-mypy]==4.2.7
pytest==7.1.3
pytest-django==4.8.0
psycopg2==2.9.9